* commit 'e967925b9112ebaea48919d867e0ff12879875ad':
  切换时间类型

+++ designer_chart/src/com/fr/design/mainframe/chart/gui/ChangeConfigPane.java
@@ -197,7 +197,7 @@ public class ChangeConfigPane extends BasicBeanPane<ChartCollection> {
         changeConfigAttr.setChangeType(configStyleButton.getSelectedIndex() == 0 ? AttrChangeType.BUTTON : AttrChangeType.CAROUSEL);
         changeConfigAttr.setStyleAttr(styleAttrPane.update());
         changeConfigAttr.setButtonColor(colorSelectBox4button.getSelectObject());
-        changeConfigAttr.setTimeInterval(timeInterval.getValue());
+        changeConfigAttr.setTimeInterval((int) timeInterval.getValue());
         changeConfigAttr.setCarouselColor(colorSelectBox4carousel.getSelectObject());
     }
